Bug and Logic changes to account for man failure conditions
Tested both runtime deploy and pre-installed for situations where one of
the mesos slaves was not configured properly and tasks failed.  We
handle this when it happens immediately by just returning the resource
to the list of requested resources and getting a new offer.  In the
future could add that slave to a list of excluded slaves.